Use shared openssl/curl libraries for building PyPi wheels so the GenomicsDB bindings can be used with TensorFlow and other libraries (#15)
* Use shared openssl/curl libraries for building PyPi wheels so the GenomicsDB bindings can be used with TensorFlow and other libraries

The base ghcr.io/genomicsdb/genomicsdb:centos_6_prereqs docker image creates static openssl/curl libraries that interfere with TensorFlow's usage of curl as they require different abi. As openssl and curl are C packages, we can just rely on openssl/curl libraries packaged with Python. For now we are just rebuilding openssl/curl with the base image. But, we should probably allow for an option in the GenomicsDB scripts to build shared libraries for openssl and curl.